Move enter_span! from util to context
