+++ context/src/lib.rs
@@ -1,2 +1,14 @@
 pub mod vulkan_render_context;
 pub mod render_context;
+
+#[macro_export]
+macro_rules! enter_span {
+    ($level:expr, $name:expr, $($fields:tt)*) => {
+        let span = tracing::span!($level, $name, $($fields)*);
+        let _enter = span.enter();
+    };
+
+    ($level:expr, $name:expr) => {
+        enter_span!($level, $name,)
+    };
+}
